[Task-3] final version
